requirements.txt
alembic==1.6.5
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
bidict==0.21.3
blinker==1.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dataclasses==0.6
dependency-injector==4.35.2
dnspython==1.16.0
flasgger==0.9.5
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SeaSurf==0.3.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
gevent==21.8.0
gevent-websocket==0.10.1
google-auth==2.3.0
greenlet==1.1.2
gunicorn==20.1.0
icalendar==4.0.7
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
packaging==21.0
parameterized==0.8.1
pdfkit==0.6.1
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
sh==1.14.2
six==1.16.0
SQLAlchemy==1.3.24
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
